Add members journey arrows
