+++ package.json
@@ -25,6 +25,8 @@
     "autoprefixer": "^10.4.14",
     "eslint-config-prettier": "^8.8.0",
     "postcss": "^8.4.23",
+    "postcss-import": "^15.1.0",
+    "postcss-nesting": "^11.2.2",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.3.0",
     "tailwindcss": "^3.3.2"
