+++ postcss.config.js
@@ -1,5 +1,8 @@
 module.exports = {
   plugins: {
+    "postcss-import": {},
+    "tailwindcss/nesting": {},
+    "postcss-import": {},
     tailwindcss: {},
     autoprefixer: {},
   },
