+++ styles/globals.css
@@ -1,3 +1,5 @@
+@import "./utils";
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
