+++ styles/utils/index.css
@@ -1,0 +1,1 @@
+@import "./clip-path.css";
